# ENH: stats.mannwhitneyu: vectorize statistic calculation