#### Reference issue
Closes gh-19642

#### What does this implement/fix?
gh-19642 asked to speed up `stats.mannwhitneyu`. The bottleneck there appeared to be computation of the statistic itself (whereas in gh-19692 the problem was with the exact p-value). This PR vectorizes the calculation of data ranks and computes the tie correction term at the same time.

#### Additional information
Speed comparison:

```python3
import numpy as np
from scipy import stats
rng = np.random.default_rng(2549824598234528)
x = rng.integers(5, size=10)
y = rng.integers(5, size=20000)
%timeit stats.mannwhitneyu(x, y)
# Before: 9.54 ms ± 12.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# After: 1.01 ms ± 3.26 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

x = rng.integers(5, size=10)
y = rng.integers(5, size=4000)
%timeit stats.mannwhitneyu(x, y)
# Before: 2.22 ms ± 602 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
# After: 421 µs ± 5.49 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

x = rng.integers(5, size=(10, 1))
y = rng.integers(5, size=(4000, 20000))
%timeit -r 1 -n 1 stats.mannwhitneyu(x, y)
# Before: 34.3 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
# After: 8.25 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)

x = rng.integers(5, size=(10, 1))
y = rng.integers(5, size=(40, 20000))
%timeit stats.mannwhitneyu(x, y)
# Before: 1.38 s ± 4.85 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# After: 64.4 ms ± 595 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```