@mdhaber 2024-01-03
Thanks for your review of gh-19776, @AtsushiSakai! This PR makes use of that vectorized `_rankdata` and the ties it returns to greatly improve the speed of `mannwhitneyu`.