@mdhaber 2024-01-03
Thanks @AtsushiSakai. Let me know about https://github.com/scipy/scipy/pull/19749#discussion_r1440146277 and I'll commit these changes.