@AtsushiSakai 2024-01-04
No other comments so far. So, let’s merge this. I think Mann-Whitney U rank test is very widely used (I’m also a heavily user for this function). So, performance improvement is very useful for a lot of users. Thanks @mdhaber 
cc @wangjiawen2013