@tengdiyouying608 2024-01-12
> #### 参考问题
> 关闭 [GH-19642型](https://github.com/scipy/scipy/issues/19642)
> 
> #### 这实现/修复了什么？
> [GH-19642](https://github.com/scipy/scipy/issues/19642) 要求加快速度.那里的瓶颈似乎是统计量本身的计算（而在[gh-19692](https://github.com/scipy/scipy/issues/19692)中，问题在于精确的p值）。此 PR 矢量化数据排名的计算，并同时计算并列校正项。`stats.mannwhitneyu`
> 
> #### 其他信息：
> 速度比较：
> 
> ```python
> import numpy as np
> from scipy import stats
> rng = np.random.default_rng(2549824598234528)
> x = rng.integers(5, size=10)
> y = rng.integers(5, size=20000)
> %timeit stats.mannwhitneyu(x, y)
> # Before: 9.54 ms ± 12.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
> # After: 1.01 ms ± 3.26 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
> 
> x = rng.integers(5, size=10)
> y = rng.integers(5, size=4000)
> %timeit stats.mannwhitneyu(x, y)
> # Before: 2.22 ms ± 602 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
> # After: 421 µs ± 5.49 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
> 
> x = rng.integers(5, size=(10, 1))
> y = rng.integers(5, size=(4000, 20000))
> %timeit -r 1 -n 1 stats.mannwhitneyu(x, y)
> # Before: 34.3 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
> # After: 8.25 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
> 
> x = rng.integers(5, size=(10, 1))
> y = rng.integers(5, size=(40, 20000))
> %timeit stats.mannwhitneyu(x, y)
> # Before: 1.38 s ± 4.85 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
> # After: 64.4 ms ± 595 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
> ```

Thank you for the update.
With same parameter as you have used, the update version and the old version cost almost same time (new ~1.4 s,old  ~1.5 s) when testing my own data, I don't know why there is no significance difference as you have observed.

```
import pandas as pd
import scipy.stats as stats
x = rng.integers(5, size=(10, 20000))
y = rng.integers(5, size=(40, 20000))
%timeit stats.mannwhitneyu(x, y)
#Before:3.04 s ± 19.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
#After:66.3 ms ± 141 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

##test my data
%timeit stats.mannwhitneyu(gene_exprs, dataset_gene)
#Before:1.48 s ± 20.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
#After:1.37 s ± 13.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
#genes_exprs is a pandas.dataframe of 30 rows x 20514 columns;dataset_gene is a pandas.dataframe of 3377 rows x 20514 columns
```